Rename helpers/helper_entity.py to helpers/helper_integration.py
