+++ homeassistant/components/switch_as_x/__init__.py
@@ -11,7 +11,7 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_ENTITY_ID
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers import device_registry as dr, entity_registry as er
-from homeassistant.helpers.helper_entity import async_handle_source_entity_changes
+from homeassistant.helpers.helper_integration import async_handle_source_entity_changes
 
 from .const import CONF_INVERT, CONF_TARGET_DOMAIN, DOMAIN
 from .light import LightSwitch
