+++ homeassistant/helpers/helper_integration.py
